+++ src/main/java/DTO/HoaDon/CreateChiTietHoaDon_DTO.java
@@ -4,4 +4,4 @@ package DTO.HoaDon;
  *
  * @author LeAnhQuan
  */
-public record CreateChiTietHoaDon_DTO(int idMonAn, int soLuong, int donGia) { }
+public record CreateChiTietHoaDon_DTO(int idMonAn, int soLuong, long donGia) { }
